# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  domain.ExamPromptItem:
    properties:
      exam_type:
        type: string
      id:
        type: string
      part:
        type: string
      prep_seconds:
        type: integer
      prompt_text:
        type: string
      speak_seconds:
        type: integer
    type: object
  domain.ExamReportData:
    properties:
      band_score:
        type: number
      cefr_level:
        type: string
      criteria: {}
      exam_type:
        type: string
      session_id:
        type: string
      status:
        type: string
    type: object
  domain.ExamSessionData:
    properties:
      band_score:
        type: number
      cefr_level:
        type: string
      exam_type:
        type: string
      session_id:
        type: string
      status:
        type: string
    type: object
...
  domain.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@phonara.app
    name: Phonara Team
  description: |-
    Backend API cho ứng dụng luyện phát âm tiếng Anh Phonara.
    Bao gồm: Authentication, Practice Sessions, Error Profile Engine,
    Content, Shadowing, Minimal Pairs, Subscription, Exam (Speechace), Daily Challenge.
  license:
    name: Proprietary
  title: Phonara API
  version: "1.0"
paths:
...
  /exam/prompts:
    get:
      description: Lấy exam prompts IELTS/TOEIC theo type và part.
      parameters:
      - description: ielts hoặc toeic
        in: query
        name: type
        type: string
      - description: part1, part2, part3…
        in: query
        name: part
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Prompts list
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ExamPromptItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Danh sách đề thi
      tags:
      - Exam
  /exam/sessions:
    get:
      description: Trả 20 session gần nhất với band_score và cefr_level (save/share,
        S33).
      produces:
      - application/json
      responses:
        "200":
          description: Sessions list
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ExamSessionData'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Lịch sử exam sessions
      tags:
      - Exam
    post:
      consumes:
      - application/json
      description: Mở session thi, kiểm tra gating Exam Pack (BR-PAY-06).
      parameters:
      - description: Prompt ID
        in: body
        name: body
        required: true
        schema:
          properties:
            prompt_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Session created
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ExamSessionData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "402":
          description: Exam Pack required
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Tạo exam session
      tags:
      - Exam
  /exam/sessions/{id}/report:
    get:
      description: Trả band_score + cefr_level + criteria sau khi Speechace đã chấm
        xong (S33).
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Exam report
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ExamReportData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Báo cáo kết quả thi
      tags:
      - Exam
  /exam/sessions/{id}/submit:
    post:
      consumes:
      - application/json
      description: Client upload audio_ref, server enqueue async Speechace scoring.
        Client KHÔNG tự chấm band (§3b.2 trust boundary).
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      - description: Audio reference
        in: body
        name: body
        required: true
        schema:
          properties:
            audio_ref:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Submitted for scoring
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ExamSessionData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Nộp bài thi (upload audio)
      tags:
      - Exam
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"